+++ CHANGELOG.md
@@ -10,6 +10,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-5356](https://openlmis.atlassian.net/browse/OLMIS-5356):
   * Extended OpenlmisResource with the getAll method.
   * Made ObjectMapper.map method synchronous.
+  * Extended OpenlmisArrayDecorator with getAllWithUniqueIds method.
 
 Improvements:
 * [OLMIS-3696](https://openlmis.atlassian.net/browse/OLMIS-3696): Added dependency and development dependency locking.